feat(location): Location 엔티티, 레포지토리 추가
