--- src/main/java/com/somemore/Location.java
@@ -1,26 +1,0 @@
-package com.somemore;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@Entity
-public class Location {
-    @Id
-    @Column(name = "id", nullable = false)
-    private Long id;
-
-    @Column(name = "address", nullable = false)
-    private String address;
-
-    @Column(name = "latitude", nullable = false)
-    private String latitude;
-
-    @Column(name = "longitude", nullable = false)
-    private String longitude;
-
-}
