+++ src/main/java/com/somemore/location/repository/LocationRepository.java
@@ -1,0 +1,8 @@
+package com.somemore.location.repository;
+
+import com.somemore.location.domain.Location;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface LocationRepository extends JpaRepository<Location, Long> {
+
+}
